:sparkles: feat(frontend): 채팅 참여 인원 ChatJointMemberCount 컴포넌트 추가 (#375)
