+++ frontend/components/ui/chat/chat-join-member-count/ChatJoinMemberCount.module.scss
@@ -1,0 +1,3 @@
+.chat_join_member_count {
+	@apply flex items-center text-xs text-customOrange;
+}
